all consersion problems are now solved
except: often level-1 is used and level is defined as size_t, so here the compiler
doesn't evaluate the covering if-statement throws a warning.


git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@267 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/data/DataVector.cpp
@@ -64,7 +64,7 @@ DataVector::DataVector(double * input, size_t size, size_t dim) : size(size), di
 
 void DataVector::resize(size_t size)
 {
-	if(size < this->size)
+	if((int)size < this->size)
 	{
 		return;
 	}
